pinax/projects/company_project/requirements.txt
--find-links=http://pypi.pinaxproject.com
--find-links=http://pypi2.pinaxproject.com

django-atomformat==0.1.1dev
creole==1.0.1

biblion>=0.1.0-dev0
django-staticfiles==0.2.0